bashrc: create global bashrc.d
In order to apply bashrc env to every user on the system we created a directory /etc/bashrc.d.
Each files in this directory will be loaded by /etc/bash.bashrc. It ensure than each user will load your file.

history.sh is created to increase the history size limit and add the date/time when you display it