awscli: replace harcoded credential path
To be able to create credential for a dedicated user, we need to have a non-harcoded path.
Remove the /root/.aws hardcoded path and replace by awscli_credentials_path. The default value still /root/.aws in order to keep the backward compatibility.
Also changed the indentation with "=" by ":" to make it easier to read.